Fixing texture issues with texture not being bound in the ctor.
Also, renaming `Buffer` to `Data` -- with `collection.mutable.Buffer`,
`java.nio.Buffer` and `AttributeBuffer`s in OpenGL,
no need to add further confusion.
