+++ src/api-opengl/scala/org/macrogl/package.scala
@@ -5,11 +5,13 @@ package org
 
 
 
-package macrogl {
+package object macrogl {
 
   case class MacroglException(msg: String) extends Exception(msg)
 
-  object Buffer {
+  type Data = java.nio.Buffer
+
+  object Data {
     type Byte = java.nio.ByteBuffer
     type Int = java.nio.IntBuffer
     type Float = java.nio.FloatBuffer
